fix(events-store): moved to separate package
