+++ python/app/templates/cart_view.jinja2
@@ -1,6 +1,5 @@
 <html>
 <body>
-{{ data }}
 <form action="/pay" method="post">
     <label for="name">Name</label>
     <input type="text" id="name" value="{{ data.name }}" readonly/><br>
